+++ axelor-helpdesk/src/main/java/com/axelor/apps/helpdesk/module/HelpdeskModule.java
@@ -24,6 +24,8 @@ import com.axelor.apps.helpdesk.db.repo.TicketRepository;
 import com.axelor.apps.helpdesk.service.MailServiceHelpDeskImpl;
 import com.axelor.apps.helpdesk.service.TicketService;
 import com.axelor.apps.helpdesk.service.TicketServiceImpl;
+import com.axelor.apps.helpdesk.service.TimerTicketService;
+import com.axelor.apps.helpdesk.service.TimerTicketServiceImpl;
 
 public class HelpdeskModule extends AxelorModule {
 
@@ -33,5 +35,6 @@ public class HelpdeskModule extends AxelorModule {
     bind(TicketRepository.class).to(TicketManagementRepository.class);
     bind(TicketService.class).to(TicketServiceImpl.class);
     bind(MailServiceBaseImpl.class).to(MailServiceHelpDeskImpl.class);
+    bind(TimerTicketService.class).to(TimerTicketServiceImpl.class);
   }
 }